+++ pkg/drift-detection/drift_evaluation_test.go
@@ -35,7 +35,7 @@ import (
 )
 
 const (
-	evaluateTimeout = 120 * 60 // high as we never want to see it
+	evaluateTimeout = int64(120 * 60) // high as we never want to see it
 )
 
 var _ = Describe("Manager: drift evaluation", func() {
